* Fixed crash that would occur in `CanvasFrame` for some video drivers  * `FFmpegFrameGrabber` now supports other input formats (devices), such as `x11grab` that can be used for screencasting  * Added `JavaCV.median()` function, and `JavaCV.fractalTriangleWave()` now respects image ROI  * Fixed background subtraction in `cvaux`  * Fixed crash inside the code for direct alignment caused by the ROI getting set outside the image plane  * Added `deltaScale` and `tryToFixPlane` to `GNImageAligner.Settings` (the first used in `ImageTransformer.Parameters` as increment, randomly selected forward or backward, for finite difference), which sometimes help to jump over local minima
git-svn-id: http://javacv.googlecode.com/svn/trunk/javacv@54 1939c828-c44a-11de-a380-73d83ba7c083

+++ src/name/audet/samuel/javacv/ProjectiveTransformer.java
@@ -322,9 +322,9 @@ public class ProjectiveTransformer implements ImageTransformer {
         }
 
         public boolean addDelta(int i) {
-            return addDelta(i, 1/*(1<<pyramidLevel)*/);
+            return addDelta(i, 1);
         }
-        public boolean addDelta(int i, int scale) {
+        public boolean addDelta(int i, double scale) {
             if (referencePoints != null && i < 8) {
                 // add one pixel..
                 projectiveParameters[i] += scale;
